+++ src/UnitTests/Machine/Observer.cs
@@ -289,7 +289,7 @@ namespace UnitTests.Machine
             Provider.Get.Heat.Heaters.Add(newHeater);
 
             Assert.AreEqual(1, numEvents);
-            Assert.AreEqual(PropertyChangeType.ObjectCollection, recordedChangeType);
+            Assert.AreEqual(PropertyChangeType.Collection, recordedChangeType);
             Assert.AreEqual(new object[] { "heat", new ItemPathNode("heaters", 0, new object[1]) }, recordedPath);
             Assert.AreSame(newHeater, recordedValue);
 
@@ -305,7 +305,7 @@ namespace UnitTests.Machine
             Provider.Get.Heat.Heaters.Add(newHeater);
 
             Assert.AreEqual(1, numEvents);
-            Assert.AreEqual(PropertyChangeType.ObjectCollection, recordedChangeType);
+            Assert.AreEqual(PropertyChangeType.Collection, recordedChangeType);
             Assert.AreEqual(new object[] { "heat", new ItemPathNode("heaters", 1, new object[2]) }, recordedPath);
             Assert.AreSame(newHeater, recordedValue);
 
@@ -335,7 +335,7 @@ namespace UnitTests.Machine
             Provider.Get.Heat.Heaters.Move(1, 0);
 
             Assert.AreEqual(2, numEvents);
-            Assert.AreEqual(PropertyChangeType.ObjectCollection, recordedChangeType);
+            Assert.AreEqual(PropertyChangeType.Collection, recordedChangeType);
             Assert.AreEqual(new object[] { "heat", new ItemPathNode("heaters", 1, new object[2]) }, recordedPath);
             Assert.AreSame(Provider.Get.Heat.Heaters[1], recordedValue);
 
@@ -351,7 +351,7 @@ namespace UnitTests.Machine
             Provider.Get.Heat.Heaters[0] = newHeater;
 
             Assert.AreEqual(1, numEvents);
-            Assert.AreEqual(PropertyChangeType.ObjectCollection, recordedChangeType);
+            Assert.AreEqual(PropertyChangeType.Collection, recordedChangeType);
             Assert.AreEqual(new object[] { "heat", new ItemPathNode("heaters", 0, new object[2]) }, recordedPath);
             Assert.AreSame(newHeater, recordedValue);
 
@@ -367,7 +367,7 @@ namespace UnitTests.Machine
             Provider.Get.Heat.Heaters.Insert(0, newHeater);
 
             Assert.AreEqual(3, numEvents);
-            Assert.AreEqual(PropertyChangeType.ObjectCollection, recordedChangeType);
+            Assert.AreEqual(PropertyChangeType.Collection, recordedChangeType);
             Assert.AreEqual(new object[] { "heat", new ItemPathNode("heaters", 2, new object[3]) }, recordedPath);
             Assert.AreEqual(Provider.Get.Heat.Heaters[2], recordedValue);
 
@@ -382,7 +382,7 @@ namespace UnitTests.Machine
             Provider.Get.Heat.Heaters.RemoveAt(0);
 
             Assert.AreEqual(3, numEvents);
-            Assert.AreEqual(PropertyChangeType.ObjectCollection, recordedChangeType);
+            Assert.AreEqual(PropertyChangeType.Collection, recordedChangeType);
             Assert.AreEqual(new object[] { "heat", new ItemPathNode("heaters", 1, new object[2]) }, recordedPath);
             Assert.AreEqual(Provider.Get.Heat.Heaters[1], recordedValue);
 
@@ -397,7 +397,7 @@ namespace UnitTests.Machine
             Provider.Get.Heat.Heaters.Clear();
 
             Assert.AreEqual(1, numEvents);
-            Assert.AreEqual(PropertyChangeType.ObjectCollection, recordedChangeType);
+            Assert.AreEqual(PropertyChangeType.Collection, recordedChangeType);
             Assert.AreEqual(new object[] { "heat", new ItemPathNode("heaters", 0, Array.Empty<object>()) }, recordedPath);
             Assert.AreEqual(null, recordedValue);